+++ src/main/java/com/scottkillen/mod/dendrology/compat/forestry/ForestryMod.java
@@ -8,11 +8,10 @@ import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.LoaderState.ModState;
 import cpw.mods.fml.common.Optional.Method;
 import cpw.mods.fml.common.event.FMLInterModComms;
-import cpw.mods.fml.common.registry.GameData;
+import cpw.mods.fml.common.registry.GameRegistry;
 import forestry.api.core.ForestryAPI;
 import forestry.api.recipes.RecipeManagers;
-import forestry.api.storage.BackpackManager;
-import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fluids.FluidRegistry;
 import net.minecraftforge.fluids.FluidStack;
@@ -22,10 +21,18 @@ public final class ForestryMod extends Integrator
     private static final String MOD_ID = "Forestry";
     private static final String MOD_NAME = MOD_ID;
 
-    private static final int FORESTER = 2;
-    private static final int BUILDER = 5;
+    private static void addBackpackItem(String backpack, ItemStack stack)
+    {
+        if (stack == null) return;
 
-    @Method(modid = MOD_ID)
+        addBackpackItem(backpack, stack.getItem(), stack.getItemDamage());
+    }
+
+    private static void addBackpackItem(String backpack, Item item, int damage)
+    {
+        sendBackpackMessage(String.format("%s@%s:%d", backpack, GameRegistry.findUniqueIdentifierFor(item), damage));
+    }
+
     private static void addBackpackItems()
     {
         Logger.forMod(TheMod.INSTANCE.modID()).info("Extending Forestry's backpacks.");
@@ -33,23 +40,22 @@ public final class ForestryMod extends Integrator
         {
             //noinspection ObjectAllocationInLoop
             final ItemStack sapling = new ItemStack(tree.saplingBlock(), 1, tree.saplingSubBlockIndex());
-            BackpackManager.backpackItems[FORESTER].add(sapling);
+            addBackpackItem("forester", sapling);
 
             //noinspection ObjectAllocationInLoop
             final ItemStack log = new ItemStack(tree.logBlock(), 1, tree.logSubBlockIndex());
-            BackpackManager.backpackItems[FORESTER].add(log);
+            addBackpackItem("forester", log);
 
             //noinspection ObjectAllocationInLoop
             final ItemStack leaves = new ItemStack(tree.leavesBlock(), 1, tree.leavesSubBlockIndex());
-            BackpackManager.backpackItems[FORESTER].add(leaves);
+            addBackpackItem("forester", leaves);
 
             //noinspection ObjectAllocationInLoop
             final ItemStack stairs = new ItemStack(tree.stairsBlock());
-            BackpackManager.backpackItems[BUILDER].add(stairs);
+            addBackpackItem("builder", stairs);
         }
     }
 
-    @Method(modid = MOD_ID)
     private static void addFarmables()
     {
         Logger.forMod(TheMod.INSTANCE.modID()).info("Adding farmable saplings to Forestry's farms.");
@@ -57,10 +63,20 @@ public final class ForestryMod extends Integrator
         {
             //noinspection ObjectAllocationInLoop
             final ItemStack sapling = new ItemStack(tree.saplingBlock(), 1, tree.saplingSubBlockIndex());
-            FMLInterModComms.sendMessage("Forestry", "add-farmable-sapling", String.format("farmArboreal@%s.%s",
-                    GameData.getBlockRegistry().getNameForObject(Block.getBlockFromItem(sapling.getItem())),
-                    sapling.getItemDamage()));
+            addFarmableSapling(sapling);
         }
+    }
+
+    private static void addFarmableSapling(ItemStack stack)
+    {
+        if (stack == null) return;
+
+        addFarmableSapling(stack.getItem(), stack.getItemDamage());
+    }
+
+    private static void addFarmableSapling(Item item, int damage)
+    {
+        sendFarmableMessage(String.format("farmArboreal@%s:%d", GameRegistry.findUniqueIdentifierFor(item), damage));
     }
 
     @Method(modid = MOD_ID)
@@ -82,10 +98,18 @@ public final class ForestryMod extends Integrator
         }
     }
 
-    @Method(modid = MOD_ID)
+    private static void sendBackpackMessage(String message)
+    {
+        FMLInterModComms.sendMessage("Forestry", "add-backpack-items", message);
+    }
+
+    private static void sendFarmableMessage(String message)
+    {
+        FMLInterModComms.sendMessage("Forestry", "add-farmable-sapling", message);
+    }
+
     private static FluidStack fluidStack(String fluidName) { return FluidRegistry.getFluidStack(fluidName, 1000); }
 
-    @Method(modid = MOD_ID)
     private static void init() { addFarmables(); }
 
     @Method(modid = MOD_ID)